wip: okay actually add all updates to commit lol

+++ Sources/Clients/RepoClient/Live.swift
@@ -46,23 +46,29 @@ extension RepoClient: DependencyKey {
 
             _ = try await databaseClient.insert(repo)
         },
-        removeRepo: { repoId in
+        updateRepo: { repo in
+            _ = try await databaseClient.update(repo)
+        },
+        deleteRepo: { repoId in
             if let repo = try? await databaseClient.fetch(.all.where(\Repo.remoteURL == repoId.rawValue)).first {
                 try await databaseClient.delete(repo)
             }
 
             await Self.downloadManager.cancelAllRepoDownloads(repoId)
         },
-        addModule: { repoId, manifest in
-            let id = manifest.id(repoID: repoId)
-            await Self.downloadManager.download(id, module: manifest)
+        installModule: { repoId, manifest in
+            Self.downloadManager.addToQueue(manifest.id(repoID: repoId), module: manifest)
         },
-        removeModule: { repoId, module in
-            let id = module.id(repoID: repoId)
+        removeModule: { id in
+            if let repo = try? await databaseClient.fetch(.all.where(\Repo.remoteURL == id.repoId.rawValue)).first,
+                  let module = repo.modules.first(where: { $0.id == id.moduleId }) {
+                try await databaseClient.delete(module)
+                try Self.fileClient.remove(fileClient.retrieveModuleDirectory(module.directory))
+            }
+
             Self.downloadManager.cancelModuleDownload(id)
-            try await databaseClient.delete(module)
-            try Self.fileClient.remove(fileClient.retrieveModuleDirectory(module.directory))
         },
+        repos: { databaseClient.observe($0) },
         downloads: {
             .init { continuation in
                 let cancellation = Self.downloadManager.states.sink { _ in
@@ -76,8 +82,7 @@ extension RepoClient: DependencyKey {
                 }
             }
         },
-        repos: { try await databaseClient.fetch($0) },
-        fetchRemoteRepoModules: { repoId in
+        fetchModulesMetadata: { repoId in
             let url = repoId.rawValue.appendingPathComponent("Manifest.json", isDirectory: false)
             let request = URLRequest(url: url)
             let (data, response) = try await URLSession.shared.data(for: request)
@@ -87,6 +92,7 @@ extension RepoClient: DependencyKey {
 }
 
 // MARK: - ModulesDownloadManager
+// FIXME: Improve download manager for queue, ect
 
 private class ModulesDownloadManager {
     let states = CurrentValueSubject<[RepoModuleID: RepoClient.RepoModuleDownloadState], Never>([:])
@@ -100,37 +106,51 @@ private class ModulesDownloadManager {
     @Dependency(\.databaseClient)
     var databaseClient
 
-    func download(_ repoModuleID: RepoModuleID, module: Module.Manifest) async {
-        guard states.value[repoModuleID]?.canRestartDownload ?? true else {
+    func addToQueue(_ repoModuleId: RepoModuleID, module: Module.Manifest) {
+        Task.detached { [weak self] in
+            await self?.download(repoModuleId, module: module)
+        }
+    }
+
+    private func download(_ repoModuleId: RepoModuleID, module: Module.Manifest) async {
+        guard states.value[repoModuleId]?.canRestartDownload ?? true else {
             return
         }
 
-        states.value[repoModuleID] = .pending
+        states.value[repoModuleId] = .pending
 
         await semaphore.wait()
         defer { semaphore.signal() }
 
-        let moduleFileURL = repoModuleID.repoId.rawValue.appendingPathComponent(module.file, isDirectory: false)
+        let moduleFileURL = repoModuleId.repoId.rawValue.appendingPathComponent(module.file, isDirectory: false)
         let request = URLRequest(url: moduleFileURL)
 
         let sequence = URLSession.shared.data(request)
-        states.value[repoModuleID] = .downloading(percent: 0)
+        states.value[repoModuleId] = .downloading(percent: 0)
 
         let task = Task<Module?, Never> {
             do {
                 for try await value in sequence {
                     switch value {
                     case let .progress(progress):
-                        states.value[repoModuleID] = .downloading(percent: progress)
+                        states.value[repoModuleId] = .downloading(percent: progress)
                     case let .value(data, response):
                         guard let response = response as? HTTPURLResponse,
-                              response.mimeType == "text/javascript",
+                              response.mimeType == "text/javascript" ||
+                                response.mimeType == "application/javascript" ||
+                                response.mimeType == "application/zip" ||
+                                response.mimeType == "application/x-zip",
                               (200..<300).contains(response.statusCode) else {
                             throw RepoClient.Error.failedToDownloadModule
                         }
 
+                        guard response.mimeType == "text/javascript" || response.mimeType == "application/javascript" else {
+                            print("Unknown mime type of file \(response.mimeType ?? "Unknown")")
+                            throw RepoClient.Error.failedToDownloadModule
+                        }
+
                         guard let directory = URL(
-                            string: "\(repoModuleID.repoId.host ?? "Default")/\(repoModuleID.moduleId.rawValue)",
+                            string: "\(repoModuleId.repoId.host ?? "Default")/\(repoModuleId.moduleId.rawValue)",
                             relativeTo: nil
                         ) else {
                             throw RepoClient.Error.failedToInstallModule
@@ -150,28 +170,28 @@ private class ModulesDownloadManager {
                 }
             } catch {
                 print(error)
-                states.value[repoModuleID] = .failed((error as? RepoClient.Error) ?? .failedToDownloadModule)
+                states.value[repoModuleId] = .failed((error as? RepoClient.Error) ?? .failedToDownloadModule)
             }
             return nil
         }
 
-        downloadTasks[repoModuleID] = task
+        downloadTasks[repoModuleId] = task
 
         guard let module = await task.value else {
-            states.value[repoModuleID] = .failed(.failedToDownloadModule)
-            downloadTasks[repoModuleID]?.cancel()
-            downloadTasks[repoModuleID] = nil
+            states.value[repoModuleId] = .failed(.failedToDownloadModule)
+            downloadTasks[repoModuleId]?.cancel()
+            downloadTasks[repoModuleId] = nil
             return
         }
 
-        states.value[repoModuleID] = .installing
+        states.value[repoModuleId] = .installing
 
-        guard var repo: Repo = try? await databaseClient.fetch(.all.where(\.remoteURL == repoModuleID.repoId.rawValue)).first else {
-            states.value[repoModuleID] = .failed(.failedToFindRepo)
+        guard var repo: Repo = try? await databaseClient.fetch(.all.where(\.remoteURL == repoModuleId.repoId.rawValue)).first else {
+            states.value[repoModuleId] = .failed(.failedToFindRepo)
             return
         }
 
-        if let index = repo.modules.firstIndex(where: { $0.id == repoModuleID.moduleId }) {
+        if let index = repo.modules.firstIndex(where: { $0.id == repoModuleId.moduleId }) {
             repo.modules.remove(at: index)
         }
 
@@ -180,16 +200,16 @@ private class ModulesDownloadManager {
         do {
             _ = try await databaseClient.update(repo)
         } catch {
-            states.value[repoModuleID] = .failed(.failedToInstallModule)
+            states.value[repoModuleId] = .failed(.failedToInstallModule)
         }
 
-        states.value[repoModuleID] = .installed
+        states.value[repoModuleId] = .installed
     }
 
-    func cancelModuleDownload(_ repoModuleID: RepoModuleID) {
-        downloadTasks[repoModuleID]?.cancel()
-        downloadTasks[repoModuleID] = nil
-        states.value[repoModuleID] = nil
+    func cancelModuleDownload(_ repoModuleId: RepoModuleID) {
+        downloadTasks[repoModuleId]?.cancel()
+        downloadTasks[repoModuleId] = nil
+        states.value[repoModuleId] = nil
     }
 
     func cancelAllRepoDownloads(_ repoId: Repo.ID) async {
